requirements.txt
PyYAML
pyzmq
numpy
tensorflow
scipy
scikit-learn
opencv-python
imageio
Pillow
dlib